--- gitlab-rb.yml
@@ -1,11 +1,0 @@
-.build_common:
-  image: ghcr.io/conecenter/c4replink:v3kc
-  variables:
-    C4COMMON_BUILDER_CMD: >
-      C4CI_BUILD_DIR=$CI_PROJECT_DIR C4REPO_MAIN_CONF=$CI_PROJECT_DIR/c4dep.ci.replink /replink.pl
-      && C4CI_BUILD_DIR=$C4COMMON_PROTO_DIR C4REPO_MAIN_CONF=$C4COMMON_PROTO_DIR/c4dep.main.replink /replink.pl
-      && C4CI_PROTO_DIR=$C4COMMON_PROTO_DIR 
-      python3 -u $C4CI_PROTO_DIR/run_with_prefix.py time 
-      python3 -u $C4CI_PROTO_DIR/build_remote.py build_common
-      --build-dir $C4CI_BUILD_DIR --push-secret $C4CI_DOCKER_CONFIG --context $CI_PROJECT_DIR 
-      --image $C4COMMON_IMAGE --commit $CI_COMMIT_SHORT_SHA --java-options "$C4BUILD_JAVA_TOOL_OPTIONS"
